Adapt tests' rtol/atol

+++ dev/tools/run_tests.py
@@ -7,5 +7,5 @@ import brian2
 
 if __name__ == '__main__':
     import numpy as np
-    if not brian2.test(float_dtype=np.float32):  # If the test fails, exit with a non-zero error code
+    if not brian2.test():  # If the test fails, exit with a non-zero error code
         sys.exit(1)
